plantFlag.sh
This file will check if a flag file exists in some directory assuming we know this directory. When file is found it will write out a string consisting of our flag to it. After it has written to file it will check permissions and change them to read only for user,group,world. Of course this is an example script of finding flag file and writing our flag to it then changing some permissions. Would like to include the chattr command to make it immutable. Please feel free to modify this as you wish... thanks!!